requirements.txt
Pillow
nose
argparse
